Only rewrite replace_re(_all) if regexp is const
Fixes #5074. This commit fixes an issue in the rewriter of
`str.replace_re` and `str.replace_re_all`. For both operators, we do two
kinds of rewrites: (1) If the first argument is a constant, then we
check whether the regular expression appears in that constant and (2) we
check whether the regular expression matches the empty string. Both of
those checks were assuming a constant regular expression but there was
no guard in place for it.  This commit adds the missing guard.