[Strings] Fix extended equality rewriter
Fixes #5090. Our extended equality rewriter was performing the following
unsound rewrite:

```
(= (str.++ Str13 Str5 Str16 Str13) (str.++ Str5 Str16 Str13 "cjyfqapanogtdznjbtqlfrmmfauwjdpvnhfpfhzsxaarlfvlvs")) ---> (and (= (str.++ Str13 Str5) Str5) (= (str.++ Str16 Str13) (str.++ Str16 Str13 "cjyfqapanogtdznjbtqlfrmmfauwjdpvnhfpfhzsxaarlfvlvs")))
```

The rule being applied was `SPLIT_EQ_STRIP_R`. The rewrite was applied
due to the following circumstances:

- The rewriter found that `(str.++ Str13 Str5)` is strictly shorter than `(str.++ Str5 Str16 Str13 "cjyfqapanogtdznjbtqlfrmmfauwjdpvnhfpfhzsxaarlfvlvs")`
- The rewriter stripped the symbolic length of the former from the
latter
- `stripSymbolicLength()` was only able to strip the first component, so
there was a remaining length of `(str.len 13)`

The rule `SPLIT_EQ_STRIP_R`, however, was implicitly assuming that the
symbolic length can either be stripped completly or not at all and was
not considering the case where only a part of the length can be
stripped.

The commit adds a flag to `stripSymbolicLength()` that makes the
function only return `true` if the whole length can be stripped from the
input. The commit also refactors the code in `stripSymbolicLength()`
slightly.

Note: It is not necessary to try to do something smart in the case where
only a partial prefix can be stripped because the rewriter tries to
apply the rule to all the different prefix combinations anyway.