Respect reordered mkScope assumptions in arith
When constructing proofs of contradictions or implications, we use
mkScope, providing it with a list of assumptions. It can re-order them.

However, arith was ignoring the reordering, and attaching the proof of
the implication/contradiction to an implication/contradiction which had
the assumptions in the original order. Now it uses the new order when
computing the implication/contradiction that it returns.